+++ inst/tinytest/test-escape.R
@@ -31,4 +31,12 @@ tab@output <- "latex"
 expect_snapshot_print(tab, "escape-issue150_caption_escape_02")
 
 
+# Escape caption and notes
+tab <- mtcars[1:3, 1:4] |>
+    setNames(c("blah_blah", "foo_bar", "c", "d")) |>
+    tt(caption = "banana_fish", notes = list("banana_fish", "b" = list(i = 1, j = 1:2, text = "banana_fish"))) |>
+    format_tt(escape = TRUE) |>
+    group_tt(j = list("foo_bar" = 1:2, "banana_fish" = 3:4))
+tab@output <- "latex"
+expect_snapshot_print(tab, "escape-issue150_caption_escape_03")